User pf act saves new img to users/uid/profile_pic
If you change your image in the user profile activity and save it, the app will upload the image using a random UUID for its name and set users/uid/profile_pic to the path of the saved image.
Problems: images aren't resized to 240x240, unused images aren't deleted from the database.