Put delete image in success listener
Before, if the write to DB was a success, the user info would update with the new img path and the old img would be deleted. This was a problem because the new image path may not be set, yet the image that it points to is deleted. That issue is now fixed.