Add shared pref for email recent change, update DB
EmailDialogFragment now sets a shared pref, KEY_EMAIL_RECENTLY_CHANGED, to true when the email is changed. MainActivity listens to this shared pref change. If it gets an update, it updates Realtime DB with the new email and updates its own email field in the sidebar (it stores the email text in a view model, created this commit). Lastly, it sets KEY_EMAIL_RECENTLY_CHANGED to false. However, I cannot test this feature right now since I need to re-authenticate. I might as well add the reauth stuff now.